+++ tests/unit/test_solvers/test_casadi_solver.py
@@ -432,6 +432,41 @@ class TestCasadiSolver(unittest.TestCase):
         ):
             solver.solve(model, t_eval)
 
+    def test_interpolant_extrapolate(self):
+        model = pybamm.lithium_ion.DFN()
+        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Chen2020)
+        experiment = pybamm.Experiment(
+            ["Charge at 1C until 4.6 V"], period="10 seconds"
+        )
+
+        param["Upper voltage cut-off [V]"] = 4.8
+
+        sim = pybamm.Simulation(
+            model,
+            parameter_values=param,
+            experiment=experiment,
+            solver=pybamm.CasadiSolver(
+                mode="safe",
+                dt_max=0.001,
+                extrap_tol=1e-3,
+                extra_options_setup={"max_num_steps": 500},
+            ),
+        )
+        with self.assertRaisesRegex(pybamm.SolverError, "interpolation bounds"):
+            sim.solve()
+
+        ci = param["Initial concentration in positive electrode [mol.m-3]"]
+        param["Initial concentration in positive electrode [mol.m-3]"] = 0.8 * ci
+
+        sim = pybamm.Simulation(
+            model,
+            parameter_values=param,
+            experiment=experiment,
+            solver=pybamm.CasadiSolver(mode="safe", dt_max=0.05),
+        )
+        with self.assertRaisesRegex(pybamm.SolverError, "interpolation bounds"):
+            sim.solve()
+
 
 class TestCasadiSolverSensitivity(unittest.TestCase):
     def test_solve_with_symbolic_input(self):